SequentialZfpCompression

Documentation for SequentialZfpCompression.

SequentialZfpCompression.CompressedArraySeq

SequentialZfpCompression.CompressedMultiFileArraySeq

Base.append!

Base.getindex

Base.ndims

Base.size

SequentialZfpCompression.SeqCompressor

SequentialZfpCompression.totalsize

SequentialZfpCompression.CompressedArraySeq — Type

CompressedArraySeq{T,Nx}

A mutable structure for storing time-dependent arrays in a compressed format.

Fields

data::Vector{UInt8}: Compressed data in byte form.

headpositions::Vector{Int64}: Positions of the beginning of each time slice in data.

tailpositions::Vector{Int64}: Positions of the end of each time slice in data.

spacedim::NTuple{Nx,Int32}: Dimensions of the spatial grid.

timedim::Int32: Number of time steps.

eltype::Type{T}: Element type of the uncompressed array.

tol::Float32: Mean absolute error that is tolerated.

precision::Float32: Controls the precision, bounding a weak relative error.

rate::Int64: Fixes the bits used per value.

SequentialZfpCompression.CompressedMultiFileArraySeq — Type

CompressedMultiFileArraySeq{T,Nx}

A compressed time-dependent array that is stored in multiple files, one per thread.

Fields

files::Vector{IOStream}: IO object for each array slice.

headpositions::Vector{Int64}: Positions of the beginning of each time slice in data.

tailpositions::Vector{Int64}: Positions of the end of each time slice in data.

spacedim::NTuple{Nx,Int32}: Dimensions of the spatial grid.

timedim::Int32: Number of time steps.

eltype::Type{T}: Element type of the uncompressed array.

tol::Float32: Mean absolute error that is tolerated.

precision::Float32: Controls the precision, bounding a weak relative error.

rate::Int64: Fixes the bits used per value.

Arguments exclusive for the constructor

filepaths::Union{Vector{String}, String}="/tmp/seqcomp": Path(s) to the files where the compressed data will be stored. If only one string is passed, the same path will be used for all threads.

Base.append! — Method

append!(compArray::CompressedArraySeq{T,N}, array::AbstractArray{T,N})

Append a new time slice to compArray, compressing array in the process.

Arguments

compArray::CompressedArraySeq{T,N}: Existing compressed array.
array::AbstractArray{T,N}: Uncompressed array to append.

```

Base.getindex — Method

getindex(compArray::AbstractCompArraySeq, timeidx::Int)

Retrieve and decompress a single time slice from compArray at timeidx.

Base.ndims — Method

ndims(compArray::AbstractCompArraySeq)

Returns the number of dimensions of the uncompressed array, including the time dimension.

Base.size — Method

size(compArray::AbstractCompArraySeq)

Returns the dimensions of the uncompressed array, with the last dimension being the time dimension.

SequentialZfpCompression.SeqCompressor — Method

SeqCompressor(dtype::DataType, spacedim::Integer...;
              rate::Int=0, tol::Real=0, precision::Real=0,
              filepaths::Union{Vector{String}, String}="",
              envVarPath::String="")

Construct a CompressedArraySeq or CompressedMultiFileArraySeq depending on the arguments.

Arguments

dtype::DataType: the type of the array to be compressed

spacedim::Integer...: the dimensions of the array to be compressed

inmemory::Bool=true: whether the compressed data will be stored in memory or in disk

rate::Int64: Fixes the bits used per value.

tol::Float32: Mean absolute error that is tolerated.

precision::Float32: Controls the precision, bounding a weak relative error.

filepaths::Union{Vector{String}, String}="": the path(s) to the files to be compressed

envVarPath::String="": the name of the environment variable that contains the path to the files to be compressed

You have the option of passing an environment variable, a file path, a vector of file paths, or nothing. If you pass a vector of file paths, the number of paths must be equal to the number of threads. If you pass a single file path, the same path will be used for all threads. If you pass an environment variable, the file path will be extracted from it. It might be useful if you are using a SLURM job scheduler, for example, since the local disk of the node can be accessed by ENV["SLURM_TMPDIR"].

Example

julia> using SequentialZfpCompression

julia> A = SeqCompressor(Float64, 4, 4)
SequentialZfpCompression.CompressedArraySeq{Float64, 2}(UInt8[], [0], [0], (4, 4), 0, Float64, 0.0f0, 0, 0)

julia> A.timedim
0

julia> size(A)
(4, 4, 0)

julia> append!(A, ones(Float64, 4, 4));

julia> A[1]
4×4 Matrix{Float64}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

julia> size(A)
(4, 4, 1)

SequentialZfpCompression.totalsize — Method

totalsize(compArray::CompressedMultiFileArraySeq)

Returns the total size of the compressed data in bytes.